Changes need to run the Mitre demo app
During Sprint 11 I spent some time trying to get the Mitre App to
run, it didn't work out of the box I had to figure out several manual
steps but ultimately it ran.

The running version of the App was turned into a container and a
manifest to run that is included in this commit.